refactor(panel): split book route responsibilities
