+++ panel/lib/features/organizations/features/realms/features/books/presentation/book/page_drag_data.dart
@@ -1,0 +1,13 @@
+part of "route.dart";
+
+class PageDrag {
+  const PageDrag({required this.pageId});
+
+  final skir.RecordId pageId;
+}
+
+class ChapterDrag {
+  const ChapterDrag({required this.chapter});
+
+  final String chapter;
+}
